fix(chat): NIP-42 challenge-bound auth, subscription leak, and code quality fixes
Critical:
- NIP-42 AUTH now POSTs challenge to /api/chat/auth-token for a freshly-signed
  event bound to the relay's challenge string, replacing pre-signed cache
- Channel switch unsubscribes old group before subscribing new, preventing
  subscription accumulation and duplicate messages

Important:
- NostrEvent.kind widened from u16 to u32 for NIP-01 forward compatibility
- CryptoService initialized once in AppState instead of from_env() per request
- WebSocket disconnected on component unmount via use_drop
- ComposeBox uses onkeydown instead of deprecated onkeypress

+++ crates/chat/src/nostr/auth.rs
@@ -131,7 +131,7 @@ impl NostrAuthService {
     pub fn sign_event_for_member(
         &self,
         encrypted_secret_key: &EncryptedBlob,
-        kind: u16,
+        kind: u32,
         content: &str,
         tags: Vec<Vec<String>>,
     ) -> Result<NostrEvent, AuthError> {
@@ -143,7 +143,7 @@ impl NostrAuthService {
         let keys =
             EventBuilder::keys_from_hex(&secret_hex).map_err(|_| AuthError::InvalidKeyData)?;
 
-        let mut builder = nostr::EventBuilder::new(nostr::Kind::Custom(kind), content);
+        let mut builder = nostr::EventBuilder::new(nostr::Kind::Custom(kind as u16), content);
 
         for tag_parts in &tags {
             if tag_parts.is_empty() {
